switched to double on data IO to avoid problems by linprog. need to output more debugging stuff for fitting taus, but getting an answer with GetTausAndLabels

+++ MatlabUtil/DataIoCypher/ReadRawData.m
@@ -5,8 +5,8 @@ function [ Time,Separation,Force ] = ...
 %   FilePath: Path to the file
     mDataSetName = '/PrhHD5GenericData';
     mData = h5read(FilePath,mDataSetName);
-    Time = mData(1,:);
-    Separation = mData(2,:);
-    Force = mData(3,:);
+    Time = double(mData(1,:));
+    Separation = double(mData(2,:));
+    Force = double(mData(3,:));
 end
